remove duplicate component
