--- src/components/Utility.js
@@ -1,13 +1,0 @@
-/*
- * check whether element is within browser viewport
- */
-export function isInViewport (elem) {
-    if (!elem) return false;
-    var bounding = elem.getBoundingClientRect();
-    return (
-        bounding.top >= 0 &&
-        bounding.left >= 0 &&
-        bounding.bottom <= (window.innerHeight || document.documentElement.clientHeight) &&
-        bounding.right <= (window.innerWidth || document.documentElement.clientWidth)
-    );
-};
