+++ src/helpers/utility.js
@@ -12,7 +12,9 @@ export function getDiffDays(dateString1, dateString2) {
     // To calculate the no. of days between two dates
     return Math.ceil(diffInTime / (1000 * 3600 * 24));
 }
-
+/*
+ * check if an image has completed loading
+ */
 export function imageOK(img) {
     if (!img) {
         return false;
@@ -28,3 +30,17 @@ export function imageOK(img) {
     }
     return true;
 }
+
+/*
+ * check whether element is within browser viewport
+ */
+export function isInViewport (elem) {
+    if (!elem) return false;
+    var bounding = elem.getBoundingClientRect();
+    return (
+        bounding.top >= 0 &&
+        bounding.left >= 0 &&
+        bounding.bottom <= (window.innerHeight || document.documentElement.clientHeight) &&
+        bounding.right <= (window.innerWidth || document.documentElement.clientWidth)
+    );
+};
